Added extra logic to match up well locations
Previously, if a sample name was identical to one from a recent day existing on the reinject list, it would get moved to Alt Inj incorrectly. Now it checks the parent well location for comparison. Reinject list that gets generated now also omits lines of empty wells, and skips over the headers in the .txt files